fix: complete ACP MCP proxy hardening
Use registered MCP callbacks, route compiled binaries through an embedded proxy entrypoint, add MCP ping support, and guard proxy calls/teardown against hangs.
