+++ packages/daemon/package.json
@@ -12,7 +12,8 @@
     "./sdk-cli-resolver": "./src/lib/agent/sdk-cli-resolver.ts",
     "./lib/internal-event-bus": "./src/lib/internal-event-bus.ts",
     "./lib/internal-command-bus": "./src/lib/internal-command-bus.ts",
-    "./lib/internal-query-bus": "./src/lib/internal-query-bus.ts"
+    "./lib/internal-query-bus": "./src/lib/internal-query-bus.ts",
+    "./lib/acp/mcp-proxy-entry": "./src/lib/acp/mcp-proxy-entry.ts"
   },
   "scripts": {
     "dev": "bun run --watch main.ts",
